feat(TargetNumber) : 대상이 되는 컴퓨터의 숫자를 저장하는 클래스
