+++ src/main/java/baseball/domain/TargetNumber.java
@@ -1,0 +1,29 @@
+package baseball.domain;
+
+import camp.nextstep.edu.missionutils.Randoms;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class TargetNumber {
+
+    private List<Integer> targetNumber;
+
+    public TargetNumber() {
+        this.targetNumber = getComputerNumbers();
+    }
+
+    private List<Integer> getComputerNumbers() {
+        List<Integer> computer = new ArrayList<>();
+
+        while (computer.size() < 3) {
+            int randomNumber = Randoms.pickNumberInRange(1, 9);
+
+            if (!computer.contains(randomNumber)) {
+                computer.add(randomNumber);
+            }
+        }
+
+        return computer;
+    }
+}
